NLP-WEB-APP-USING-FLASK/requirements.txt
flask==2.0.1
paralleldots==1.0.5
python-dotenv==0.19.0
